[lldb] Factor out iteration over runtime types from GetChildCompilerTypeAtIndex()
This patch introduces a new class SwiftRuntimeTypeVisitor that exists
to unify iteration over runtime type information. The visitor callback
has closure parameters that can be called to make additional expensive
queries on a child.

TODO: This is not the final evolution step.

- We probably should remove the "depth" parameter entirely and implement the access path computation for GetIndexOfChildMemberWithName at a different layer.

- We could cache the results for the last execution context.

Relanding with an off-by-one error fixed in
GetExistentialSyntheticChildren that was caught by ASAN!