[lldb][swift] Change unwind heuristic for Q funclets in arm64e
With arm64e, many more branches are generated, which cause the debugger
to stop more often while stepping. Among those stops are code regions
where the debugger gets confused about what is stored in the async
register stack slot.

This commit revives a heuristic that was previously used by default for
all architectures and that works in all PC addresses, but that
sacrifices unwinding o recursive functions. It is only used if the
target is arm64e.