Don't stop on breakpoints when running the swift Object Description function (#10693)
* The Swift `object description` expression was being run without
setting "IgnoreBreakpointsInExpressions", so it was stopping on
breakpoints.  That's not how the ObjC object description is run,
and is not what people expect here.

I added the setting to the options in the Swift case, and added a
test that we don't stop on breakpoints in the Object Description
expression.

* Mark the tests as @swiftTest and fix the test name.