[lldb] Reduce max-children-depth default for readability (#10683)
* [lldb] Reduce max-children-depth default for readability

Decrease `target.max-children-depth` from 6 to 4. Often the value of 6 produces too much
output, particularly with Swift data where, unlike ObjC, classes are not treated as
pointers and thus are traversed just like structs.

For a hypothetical recursive struct with just two members/properties, a depth of 6
results in 63 (2^6-1) child values printed. In practice, data types commonly have more
than two fields, and so hundred of child values can be printed. An output of only 63
would be closer to a best case scenario. This doesn't factor in lldb's data output
syntax, which adds additional non-data lines to the output.

When all children must be shown, frame variable and expression both support the -A
(--show-all-children) flag.

rdar://145327522

* Fix tests for max-children-depth

* Fix test/Shell/SymbolFile/PDB/udt-layout.test