[clang][cas] Handle CAS errors from OutputFile::keep()
When finishing output files we need to handle all possible Error kinds
or we will have undefined behaviour. In particular, if storing to the
CAS failed we could silently fail to add the main output leading to
mysterious downstream failures if we tried to replay that compilation,
or for modules when we tried to load the pcm contents.

rdar://151077068
(cherry picked from commit f2b5016c1e0d18799e506d796877e7b78815d515)