[lldb][NFCI] Implement language-agnostic diagnostic for non-captured vars in "frame var"
This commit adds the language-agnostic logic to support diagnosing when
a `frame var` command failed because a variable is not visible in the
current frame, but the current frame is a "closure-like"(*) function and
it could have access to the variable, had it been captured by the
closure.

(*) A lambda in C++, a block in C or Objective C, a closure in Swift.

The logic relies on language plugins being able to produce the name of
the function where the current closure is defined, given the mangled
name of the closure. Since this commit does not implement support for
this in any languages, it is NFCI. It contains all the necessary pieces
to upstream the code, once we implement Language support for at least
one upstream language.