[lldb] Make private type reflection metadata available from the file cache
This fixes a bug in LLDB's file cache. Private types have anonymous
context descriptors, which are just a string that is the metadata
pointer in hex. This is a problem when the meatdata was read from a
file because then the address isn't comparable against the same type
constructed from in-process memory. This patch fixes this by
implementing a new hook to canonicalize file addresses to in-process
addresses in MemoryReader.

rdar://151330005