[lldb] Add missing return type annotations
In a thorny corner of the language C++ doesn't actually type check the
return type of a function passed into a std::function argument. Here
the type inference in the lambda resulted in a ChildInfo being
returned where an llvm::Expected<ChildInfo> was expected, which
through coincidence works very often, but not always, leading to a
crash.

rdar://153137542