[lldb] Initialize a shared, blank SwiftASTContext in Objective-C frames
The result of an Objective-C expression can be a Swift
ValueObject. This means that `po foo` in Objective-C may run a Swift
expression to print the object description of the result of the
Objective-C expression `foo`. That expression has very few
dependencies:

```
Swift._DebuggerSupport.stringForPrintObject(Swift.unsafeBitCast(0x001234, to: AnyObject.self))
```

To avoid scanning potentially incompatible compiler flags and/or doing
implicit imports in anotherwise explicitly built project, this patch
ensures that one SwiftASTContext per triple is created, with no other
contextual information. This makes `po` in ObjC much faster and
reliable, the downside of this is that

```expr -l Swift -- import ModuleName```

will only work reliably for SDK modules outside of Swift frames.

rdar://153332037