[lldb][swift] Cache and lazily-evaluate Task names in OperatingSystemSwiftTasks
Computing a Task name can be an expensive operation, especially over
slower connections to the inferior process. To improve this situation,
this patch:

* Makes the computation lazy, and
* Caches the result of the computation (a task name is immutable).

The first point is the most important, as prior to this we were
computing the task name on every stop, even when the same MemoryThread
was reused between two consecutive stops.

On local testing with a wired connection, this reduces stepping time by
around 20-30ms, out of an average of 230ms.