CAS validation improvements (#10966)
* [CAS] Validate upstream CAS storage

When asked to validate the CAS, also validate the upstream CAS if
present. The entire upstream CAS is available as backup storage of the
currently active CAS. If the upstream CAS is corrupted, it can propagate
the error into downstream error, thus when upstream CAS is invalid, the
entire UnifiedCAS should be invalid.

* [CAS] Do not check if action cache value exists in the object store

Fix a false positive/crash when validating action cache from unified
data base by stop checking the cached value exists in the object store.
First of all, there is a bug that the check is always done against
primary object store, where the object can be missing. It is also
possible that the CAS is in a state where the cache keys are fault in,
but not the objects the cache pointed to. There is no good way to
enforce it, thus the check is too strict.

rdar://155231696
(cherry picked from commit 09bbceb22f43a3bf474f66b4c22266f7bfd0ff30)