[lldb] Fix caching, add lookup failures to TSSTyperef's LookupClangType
By profiling LLDB running frame variable in a large application with
many many .o files, I noticed that one of the biggest bottlenecks of the
operation was looking (and failing) to find a certain potential clang
type over and over again.

While fixing this I found that we will cache the result, but not the
decl context to find that result, which introduces the risk of
returning the wrong type if the type is cached, but a different decl
context is used.