[clang][DebugInfo] Handle empty comp_dir in CodeGen
Fix a logic mistake in DebugInfo CodeGen that if the compilation
directory is empty, the leading separator is dropped with emitting path.
This code path is usually not possible in current state unless the
compilation is done from a file system that do not have current working
directory and compilation caching include tree file system is such a
file system. This causes PCMs are encoded with wrong path in the
gmodule compilation, causing lost of debug info.

Note this is a reduced and safer version from upstream fix:
https://github.com/llvm/llvm-project/pull/150130

The test is in swift repo since the only trigger in current branch is to
trigger from swift caching builds.

rdar://156759645