[lldb] Skip prologue when stepping through swift_task_switch
Plans stepping through a swift_task_switch are often the last sub-plan
for a StepOver action, meaning the plan's destination is also the PC
where users will take control of the program. As such, it is crucial
that these plans skip over the prologue of their destination function,
otherwise most variables won't be in scope.

Because x86 and arm diverge on the codegen for `}` in a scope, I had to
add a `print` statement in the for loop of the affected test.

(cherry picked from commit 3ed5f1d9f39740f7e4bf2444db9ec6b6cc75e6e7)